+++ changelog.md
@@ -31,6 +31,7 @@
     - Added `min_value` and `max_value` attribute for `number`
     - Added `x_must_be = {"key": "{label}", "value": "..."}` attribute to show field when a certain field has a certain value
 22. Added data saver mode
+23. Improved color picker
 
 ## v3.0.0